Packing list refactor

+++ app/api/category.py
@@ -3,7 +3,7 @@ from fastapi_sqlalchemy import db
 from pydantic import BaseModel
 from sqlalchemy import or_
 
-from models.base import User, Category, Item
+from models.base import User, Category, Item, ItemCategory
 from utils.auth import authenticate
 
 route = APIRouter()
@@ -11,12 +11,12 @@ route = APIRouter()
 
 class CategoryType(BaseModel):
     name: str
-    consumable: bool = None
+    consumable: bool = False
 
 
 @route.post("")
 def create(payload: CategoryType, user: User = Depends(authenticate)):
-    new_category = Category(user_id=user.id, **payload.dict())
+    new_category = Category(user_id=user.id, name=payload.name)
 
     try:
         db.session.add(new_category)
@@ -25,7 +25,19 @@ def create(payload: CategoryType, user: User = Depends(authenticate)):
     except:
         raise HTTPException(400, "Unable to create category.")
 
-    return new_category
+    position = db.session.query(
+        ItemCategory).filter_by(user_id=user.id).count()
+    item_category = ItemCategory(
+        category_id=new_category.id, user_id=user.id, sort_order=position)
+
+    try:
+        db.session.add(item_category)
+        db.session.commit()
+        db.session.refresh(item_category)
+    except:
+        raise HTTPException(400, "Unable to create item category.")
+
+    return item_category
 
 
 @route.get("")
@@ -35,12 +47,12 @@ def fetch(user: User = Depends(authenticate)):
 
 class CategoryUpdateType(BaseModel):
     name: str = None
-    consumable: bool = None
 
 
 @route.put("/{category_id}")
 def update(category_id, payload: CategoryUpdateType, user: User = Depends(authenticate)):
-    category = db.session.query(Category).filter_by(id=category_id, user_id=user.id).first()
+    category = db.session.query(Category).filter_by(
+        id=category_id, user_id=user.id).first()
     if not category:
         raise HTTPException(400, "Category does not exist.")
 
@@ -59,24 +71,25 @@ def update(category_id, payload: CategoryUpdateType, user: User = Depends(authen
 
 @route.delete("/{category_id}")
 def delete(category_id, user: User = Depends(authenticate)):
-    category = db.session.query(Category).filter_by(id=category_id, user_id=user.id).first()
+    category = db.session.query(Category).filter_by(
+        id=category_id, user_id=user.id).first()
     if not category:
         raise HTTPException(400, "Category does not exist.")
 
-    # Reassign associated items
-    category_items = db.session.query(Item).filter_by(category_id=category.id, user_id=user.id).all()
-    for item in category_items:
-        item.category_id = None
+    # Uncategorize associated items
+    item_category = db.session.query(ItemCategory).filter_by(
+        category_id=category.id, user_id=user.id).first()
 
-    try:
-        db.session.commit()
-    except Exception:
-        raise HTTPException(400, "Unable to reassign existing items.")
+    if item_category:
+        items = db.session.query(Item).filter_by(category_id=item_category, user_id=user.id).all()
+        for item in items:
+            item.category_id = None
 
-    try:
-        db.session.delete(category)
-        db.session.commit()
-    except Exception:
-        raise HTTPException(400, "Unable to delete category.")
+        try:
+            db.session.delete(item_category)
+            db.session.delete(category)
+            db.session.commit()
+        except Exception:
+            raise HTTPException(400, "Unable to delete category.")
 
     return True
